feat(leave_request): add API for leave request

+++ src/main/java/com/lamdayne/humify/attendance/dto/request/CreateLeaveTypeRequest.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CreateLeaveTypeRequest {
@@ -19,7 +21,7 @@ public class CreateLeaveTypeRequest {
     @Pattern(regexp = "^[A-Z_]+$", message = "LEAVE_TYPE_CODE_INVALID")
     private String code;
 
-    private Integer maxDays;
+    private BigDecimal maxDays;
 
     private Boolean isPaid;
